requirements.txt
tweepy
PyYAML
boto3
requests-aws4auth
elasticsearch
